Refactor model builders into their own package
